requirements.txt
flask
flask-cors
dnspython
